requirements.txt
amqp==2.5.2
appnope==0.1.0
backcall==0.1.0
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
django-cors-headers==3.2.1
django-crontab==0.7.1
django-filter==2.2.0
django-haystack==2.8.1
django-redis==4.11.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
elasticsearch==2.4.1
fdfs-client-py==1.2.6
idna==2.9
importlib-metadata==1.6.0
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
kombu==4.6.8
mutagen==1.44.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycryptodomex==3.9.7
Pygments==2.6.1
PyJWT==1.7.1
PyMySQL==0.9.3
python-alipay-sdk==1.8.0
pytz==2019.3
QQLoginTool==0.3.0
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
vine==1.3.0
wcwidth==0.1.9
zipp==3.1.0
